Move install tests to the same file.
